+++ src/main.rs
@@ -31,8 +31,8 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     Ok(())
 }
 
-fn tki(width: usize, height: usize) -> board::Board {
-    let mut board = board::Board::new(width, height);
+fn tki(width: u32, height: u32) -> board::Board {
+    let mut board = board::Board::new(width as usize, height as usize);
     board.place(&piece::Piece::O(piece::Cell { x: 0, y: 0 }, piece::Orientation::NORTH)).unwrap();
     board.place(&piece::Piece::I(piece::Cell { x: 4, y: 0 }, piece::Orientation::NORTH)).unwrap();
     board.place(&piece::Piece::Z(piece::Cell { x: 2, y: 1 }, piece::Orientation::EAST)).unwrap();
